More fixes

+++ korge-core/src/commonTest/kotlin/korlibs/math/geom/bezier/AssertExt.kt
@@ -1,7 +1,7 @@
 package korlibs.math.geom.bezier
 
+import korlibs.math.*
 import korlibs.math.geom.*
-import korlibs.math.isAlmostEquals
 
 fun <T : Any> assertEqualsFloat(
     expected: T?,
@@ -27,7 +27,7 @@ private fun <T : Any> T?.isAlmostEqualsGeneric(
     val e = this
     if (e == null || a == null) return (e == null) && (a == null)
     return when (e) {
-        is Point -> e.isAlmostEquals((a as? Point?) ?: return false, absoluteTolerance.toFloat())
+        is Point -> e.isAlmostEquals((a as? Point?) ?: return false, absoluteTolerance)
         is MPoint -> e.isAlmostEquals((a as? MPoint?) ?: return false, absoluteTolerance)
         is Float -> {
             if (a !is Float?) return false
