+++ korge-foundation/src/commonTest/kotlin/korlibs/math/geom/AssertExt.kt
@@ -1,8 +1,8 @@
 package korlibs.math.geom
 
 import korlibs.datastructure.*
+import korlibs.math.*
 import korlibs.math.geom.bezier.*
-import korlibs.math.isAlmostEquals
 import kotlin.test.*
 
 fun <
@@ -32,15 +32,15 @@ private fun <T : Any> T?.isAlmostEqualsGeneric(
     val e = this
     if (e == null || a == null) return (e == null) && (a == null)
     return when (e) {
-        is Ray -> e.isAlmostEquals((a as? Ray?) ?: return false, absoluteTolerance.toFloat())
-        is Point -> e.isAlmostEquals((a as? Point?) ?: return false, absoluteTolerance.toFloat())
+        is Ray -> e.isAlmostEquals((a as? Ray?) ?: return false, absoluteTolerance)
+        is Point -> e.isAlmostEquals((a as? Point?) ?: return false, absoluteTolerance)
         is Vector3F -> e.isAlmostEquals((a as? Vector3F?) ?: return false, absoluteTolerance.toFloat())
         is Matrix3 -> e.isAlmostEquals((a as? Matrix3?) ?: return false, absoluteTolerance.toFloat())
         is Matrix4 -> e.isAlmostEquals((a as? Matrix4?) ?: return false, absoluteTolerance.toFloat())
         is Quaternion -> e.isAlmostEquals((a as? Quaternion?) ?: return false, absoluteTolerance.toFloat())
         is EulerRotation -> e.isAlmostEquals((a as? EulerRotation?) ?: return false, absoluteTolerance.toFloat())
         is MPoint -> e.isAlmostEquals((a as? MPoint?) ?: return false, absoluteTolerance)
-        is Angle -> e.isAlmostEquals((a as? Angle?) ?: return false, absoluteTolerance.toFloat())
+        is Angle -> e.isAlmostEquals((a as? Angle?) ?: return false, absoluteTolerance)
         is Float -> {
             if (a !is Float?) return false
             if (e.isNaN() && a.isNaN()) return true
@@ -51,12 +51,12 @@ private fun <T : Any> T?.isAlmostEqualsGeneric(
             if (e.isNaN() && a.isNaN()) return true
             e.isAlmostEquals(a, absoluteTolerance)
         }
-        is Matrix -> e.isAlmostEquals((a as Matrix), absoluteTolerance.toFloat())
-        is MatrixTransform -> e.isAlmostEquals((a as MatrixTransform), absoluteTolerance.toFloat())
-        is Rectangle -> e.isAlmostEquals((a as Rectangle), absoluteTolerance.toFloat())
+        is Matrix -> e.isAlmostEquals((a as Matrix), absoluteTolerance)
+        is MatrixTransform -> e.isAlmostEquals((a as MatrixTransform), absoluteTolerance)
+        is Rectangle -> e.isAlmostEquals((a as Rectangle), absoluteTolerance)
         is MRectangle -> e.isAlmostEquals((a as MRectangle), absoluteTolerance)
         is Bezier -> e.points.isAlmostEqualsGeneric((a as? Bezier)?.points, absoluteTolerance)
-        is Bezier.ProjectedPoint -> e.isAlmostEquals((a as Bezier.ProjectedPoint), absoluteTolerance.toFloat())
+        is Bezier.ProjectedPoint -> e.isAlmostEquals((a as Bezier.ProjectedPoint), absoluteTolerance)
         is PointList -> e.toList().isAlmostEqualsGeneric((a as? PointList)?.toList(), absoluteTolerance)
         is FloatArrayList -> e.toList().isAlmostEqualsGeneric((a as? FloatArrayList)?.toList(), absoluteTolerance)
         is DoubleArrayList -> e.toList().isAlmostEqualsGeneric((a as? DoubleArrayList)?.toList(), absoluteTolerance)
