+++ korge/src/commonMain/kotlin/korlibs/korge/view/filter/ColorMatrixFilter.kt
@@ -16,7 +16,7 @@ import korlibs.math.geom.*
  * - [ColorMatrixFilter.GRAYSCALE_MATRIX] - Used to make the colors grey
  * - [ColorMatrixFilter.IDENTITY_MATRIX]  - Doesn't modify the colors at all
  */
-class ColorMatrixFilter(colorMatrix: Matrix4, blendRatio: Float = 1f) : ShaderFilter() {
+class ColorMatrixFilter(colorMatrix: Matrix4, blendRatio: Double = 1.0) : ShaderFilter() {
     object ColorMatrixUB : UniformBlock(fixedLocation = 5) {
         val u_ColorMatrix by mat4()
         val u_BlendRatio by float()
@@ -69,7 +69,7 @@ class ColorMatrixFilter(colorMatrix: Matrix4, blendRatio: Float = 1f) : ShaderFi
      * - Values between [0 and 1] would be an interpolation between those colors.
      * */
     @ViewProperty
-	var blendRatio: Float = blendRatio
+	var blendRatio: Double = blendRatio
 
     override val programProvider: ProgramProvider get() = ColorMatrixFilter
